src/requirements.txt
discord.py[voice]
aiohttp[speedups]
pyyaml
sympy
googletrans